backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ansible-tool-backend"
version = "0.1.0"
description = "Internal enterprise web application for Ansible-based network automation"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "ansible-runner>=2.4.0",
  "celery[redis]>=5.4.0",
  "croniter>=2.0.5",
  "cryptography>=42.0.8",
  "fastapi>=0.115.0",
  "jinja2>=3.1.4",
  "ldap3>=2.9.1",
  "openpyxl>=3.1.5",
  "pandas>=2.2.2",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.4.0",
  "python-dateutil>=2.9.0.post0",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.34",
  "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
